Fix frameskip reporting in Libretro
If we call video_cb with NULL, it registers a completed a frame and thus the FPS display will still show <=60fps regardless of our frameskip settings in gpsp core options.  This change makes it show a value I'd more expect (i.e. frameskip set to 1 will show <= 30fps) but I'm not sure if it's necessarily the correct way to achieve this.