Restore previous frameskip options code
I accidentally uploaded a test version of the file where I hardcoded the frameskip to run at 30fps.

The Auto frameskip uses the audio buffer to determine whether or not we need to skip, and does some floating point calculations which I don't think are required (converting them to Integers doesn't seem to affect sound quality on my low-end device).  So could look at that in future.